Added dedicated database.py file including implementation and separation of the DB logic from the sever.py
There is a new class including connection_management that uses pattern matching to check various connection DB_name and verify which works. This makes a more robust solution in case there are various DB options inside of the project - the tool just tries a valid connection string and keep (cache) the connection that worked.